+++ include/boost/beast/core/basic_stream.hpp
@@ -498,7 +498,7 @@ public:
         @return A copy of the executor that stream will use to dispatch handlers.
     */
     executor_type
-    get_executor() const noexcept
+    get_executor() noexcept
     {
         return impl_->ex();
     }