+++ test/beast/core/stream_traits.cpp
@@ -213,14 +213,14 @@ public:
 
     struct async_read_stream
     {
-        net::io_context::executor_type get_executor();
+        net::io_context::executor_type get_executor() noexcept;
         template<class MutableBufferSequence, class ReadHandler>
         void async_read_some(MutableBufferSequence const&, ReadHandler&&);
     };
 
     struct async_write_stream
     {
-        net::io_context::executor_type get_executor();
+        net::io_context::executor_type get_executor() noexcept;
         template<class ConstBufferSequence, class WriteHandler>
         void async_write_some(ConstBufferSequence const&, WriteHandler&&);
     };
@@ -231,7 +231,7 @@ public:
 
     struct async_stream : async_read_stream, async_write_stream
     {
-        net::io_context::executor_type get_executor();
+        net::io_context::executor_type get_executor() noexcept;
         template<class MutableBufferSequence, class ReadHandler>
         void async_read_some(MutableBufferSequence const&, ReadHandler&&);
         template<class ConstBufferSequence, class WriteHandler>