+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 253:
 
 * Fix async_detect_ssl handler type
+* member get_executor const-correctness
 
 --------------------------------------------------------------------------------